Rangos de Traslapes terminados

+++ src/Proceso/Difusificacion.java
@@ -32,10 +32,10 @@ public class Difusificacion {
         return y_grado_membresia;
     }
     double y_grado_membresia;
-    public double[][] calculaT(ArrayList<Competencia> puntosCC) {
-        double[][] puntosC = new double[8][2];
-        double[][] puntosCO = new double[8][2];
-        double[][] traspales = new double[8][2];
+    public int[][] calculaT(ArrayList<Competencia> puntosCC) {
+        int[][] puntosC = new int[8][2];
+        int[][] puntosCO = new int[8][2];
+        int[][] traspales = new int[8][2];
         double puntoi, puntof;
         int F = 0, C = 0;
         //double traslape1, traslape2, traslape3, traslape4;
@@ -73,8 +73,8 @@ public class Difusificacion {
          * -----------
          * */
         for (int h = 0; h < puntosCO.length; h++) {
-            double resta = puntosCO[h][1] - puntosCO[h][0];
-            double porcentaje = resta * 0.75;
+            int resta = puntosCO[h][1] - puntosCO[h][0];
+            int porcentaje = (int)(resta * 0.75);
             traspales[h][0] = puntosCO[h][1] - porcentaje;
             traspales[h][1] = puntosCO[h][0] + porcentaje;
         }
